+++ Sources/Mock.swift
@@ -79,8 +79,11 @@ public struct Mock: Equatable {
     /// Add a delay to a certain mock, which makes the response returned later.
     public var delay: DispatchTimeInterval?
 
-    /// The callback which will be executed everytime this `Mock` was used. Can be used within unit tests for validating that a request has been executed.
+    /// The callback which will be executed everytime this `Mock` was completed. Can be used within unit tests for validating that a request has been executed.
     public var completion: (() -> Void)?
+
+    /// The callback which will be executed everytime this `Mock` was started. Can be used within unit tests for validating that a request has been started.
+    public var onRequest: (() -> Void)?
     
     private init(url: URL? = nil, ignoreQuery: Bool = false, dataType: DataType, statusCode: Int, data: [HTTPMethod: Data], additionalHeaders: [String: String] = [:], fileExtensions: [String]? = nil) {
         self.url = url ?? URL(string: "https://mocked.wetransfer.com/\(dataType.rawValue)/\(statusCode)/")!