+++ Changelog.md
@@ -1,7 +1,8 @@
 ## Changelog
 
 ### Next
-- A new completion callback can be set on `Mock` to use for expectation fulfilling once a `Mock` is used.
+- A new completion callback can be set on `Mock` to use for expectation fulfilling once a `Mock` is completed.
+- A new onRequest callback can be set on `Mock` to use for expectation fulfilling once a `Mock` is requested.
 - Updated to Swift 5.0
 - Only dispatch to the background queue if needed
 - Correctly handle cancellation of delayed responses